Remove obsolete `RocketChatAnnouncement` API
