--- apps/meteor/app/lib/client/lib/index.js
@@ -1,2 +1,0 @@
-export { RocketChatAnnouncement } from './RocketChatAnnouncement';
-export { LoginPresence } from './LoginPresence';
