+++ apps/meteor/app/lib/client/lib/index.ts
@@ -1,0 +1,1 @@
+export { LoginPresence } from './LoginPresence';
